Done some improvements to pages functionality

+++ src/main/java/pl/coderslab/controller/CategoryController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import pl.coderslab.entity.Category;
 import pl.coderslab.repository.CategoryRepository;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.List;
 
@@ -21,14 +22,28 @@ public class CategoryController {
     private CategoryRepository categoryRepository;
 
     @RequestMapping(value = "/add", method = RequestMethod.GET)
-    public String addGet(Model model){
+    public String addGet(Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         model.addAttribute("addCategory", "addCategory");
         model.addAttribute("category", new Category());
         return "home";
     }
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
-    public String addPost(@Valid Category category, BindingResult result, Model model){
+    public String addPost(@Valid Category category, BindingResult result, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         if(result.hasErrors()){
             model.addAttribute("addCategory", "addCategory");
             return "home";
@@ -38,14 +53,28 @@ public class CategoryController {
     }
 
     @RequestMapping(value = "/edit/{id}", method = RequestMethod.GET)
-    public String editGet(@PathVariable("id") Long id, Model model){
+    public String editGet(@PathVariable("id") Long id, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         model.addAttribute("editCategory", "editCategory");
         model.addAttribute("category", categoryRepository.findById(id));
         return "home";
     }
 
     @RequestMapping(value = "/edit/*", method = RequestMethod.POST)
-    public String editPost(@Valid Category category, BindingResult result, Model model){
+    public String editPost(@Valid Category category, BindingResult result, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         if(result.hasErrors()){
             model.addAttribute("editCategory", "editCategory");
             return "home";
@@ -55,13 +84,27 @@ public class CategoryController {
     }
 
     @RequestMapping(value = "/all", method = RequestMethod.GET)
-    public String all(Model model){
+    public String all(Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         model.addAttribute("allCategories", "allCategories");
         return "home";
     }
 
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
-    public String delete(@PathVariable("id") Long id){
+    public String delete(@PathVariable("id") Long id, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         categoryRepository.deleteById(id);
         return "home";
     }
