+++ src/main/java/pl/coderslab/controller/ProductController.java
@@ -10,6 +10,7 @@ import pl.coderslab.entity.Product;
 import pl.coderslab.repository.CategoryRepository;
 import pl.coderslab.repository.ProductRepository;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.List;
 
@@ -23,7 +24,14 @@ public class ProductController {
     private CategoryRepository categoryRepository;
 
     @RequestMapping(value = "/add", method = RequestMethod.GET)
-    public String addGet(Model model){
+    public String addGet(Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         model.addAttribute("product", new Product());
         model.addAttribute("categories", allCategories());
         model.addAttribute("addProduct", "addProduct");
@@ -31,7 +39,14 @@ public class ProductController {
     }
 
     @RequestMapping(value = "/add", method = RequestMethod.POST)
-    public String addPost(@Valid Product product, BindingResult result, Model model){
+    public String addPost(@Valid Product product, BindingResult result, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         if(result.hasErrors()){
             model.addAttribute("addProduct", "addProduct");
             return "home";
@@ -47,14 +62,28 @@ public class ProductController {
     }
 
     @RequestMapping(value = "/edit/{id}", method = RequestMethod.GET)
-    public String editGet(@PathVariable("id") Long id, Model model){
+    public String editGet(@PathVariable("id") Long id, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         model.addAttribute("editProduct", "editProduct");
         model.addAttribute("product", productRepository.findTopById(id));
         return "home";
     }
 
     @RequestMapping(value = "/edit/*", method = RequestMethod.POST)
-    public String editPost(@Valid Product product, BindingResult result, Model model){
+    public String editPost(@Valid Product product, BindingResult result, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         if(result.hasErrors()){
             model.addAttribute("editProduct", "editProduct");
             return "home";
@@ -64,21 +93,42 @@ public class ProductController {
     }
 
     @RequestMapping(value = "/search", method = RequestMethod.POST)
-    public String search(@RequestParam("name") String name, Model model){
+    public String search(@RequestParam("name") String name, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         model.addAttribute("allProducts", "allProducts");
         model.addAttribute("products", productRepository.findByNameContaining(name));
         return "home";
     }
 
     @RequestMapping(value = "/all", method = RequestMethod.GET)
-    public String all(Model model){
+    public String all(Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         model.addAttribute("allProducts", "allProducts");
         model.addAttribute("products", allProducts());
         return "home";
     }
 
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.GET)
-    public String delete(@PathVariable("id") Long id){
+    public String delete(@PathVariable("id") Long id, Model model, HttpSession session){
+        Object object = session.getAttribute("user");
+        if(object == null){
+            model.addAttribute("logged", null);
+            model.addAttribute("loginForm", "loginForm");
+            return "home";
+        }
+        model.addAttribute("logged", "logged");
         productRepository.deleteById(id);
         return "home";
     }
